Math Quiz
Riya participated in a math quiz competition. At the end of the contest, she had to solve a tricky question to win the first prize.
She has been given a few numbers, M multiple ranges and her task is to add all the numbers in each corresponding range. Help Riya to solve this question.
Write a program that reads the space-separated numbers and M multiple ranges and prints the sum of all the numbers in each range.